implement --hdrsrc substitution mechanism
this will cause rcb to scan the substituted header path when a
matching #include <*> directive is found.
this allows for example to install libulz into the system directories
/include/ulz, /lib, so regular library users can include <ulz/strlib.h>
instead of "../lib/include/strlib.h" and just link with -lulz.
those headers tho will contain rcb tags that point to a non-existing
source directory.
rcb can still be used in such a case by using
--hdrsrc=ulz:../lib/include
so <ulz/strlib.h> will be transformed into "../lib/include/strlib.h"
and the rcb tags scanned. inclusions with <> are otherwise ignored.
the separator is between the expression to match and the replacement
string is ":".